conformerRmsd: fix size_t coordinate offset handling, binary search, and overflow guards
- Remove dead `disc` variable in symmetricEigenvalues3x3; near-degenerate
  inputs are handled by the existing fmax guards on sqrtPP and acos argument
- Promote coordOffsets to size_t throughout: PinnedHostVector, host packing
  index (base), AsyncDeviceVector, and kernel parameter — prevents silent
  int overflow for large batches (>INT_MAX coordinate elements)
- Replace O(numMols) linear scan with O(log numMols) binary search in
  conformerRmsdBatchKernel for molecule lookup
- Guard cumulative pairOffsets prefix sum against int overflow using int64_t
  running total before narrowing back to int

+++ src/conformer_rmsd.h
@@ -69,16 +69,16 @@ void conformerRmsdMatrixGpu(cuda::std::span<const double> coords,
  * @param prealigned      If true, skip Kabsch alignment.
  * @param stream          CUDA stream.
  */
-void conformerRmsdBatchMatrixGpu(cuda::std::span<const double> coords,
-                                  cuda::std::span<double*>      rmsdOutputs,
-                                  cuda::std::span<const int>    pairOffsets,
-                                  cuda::std::span<const int>    coordOffsets,
-                                  cuda::std::span<const int>    numConfsPerMol,
-                                  cuda::std::span<const int>    numAtomsPerMol,
-                                  int                           numMols,
-                                  int                           totalPairs,
-                                  bool                          prealigned,
-                                  cudaStream_t                  stream);
+void conformerRmsdBatchMatrixGpu(cuda::std::span<const double>  coords,
+                                  cuda::std::span<double*>       rmsdOutputs,
+                                  cuda::std::span<const int>     pairOffsets,
+                                  cuda::std::span<const size_t>  coordOffsets,
+                                  cuda::std::span<const int>     numConfsPerMol,
+                                  cuda::std::span<const int>     numAtomsPerMol,
+                                  int                            numMols,
+                                  int                            totalPairs,
+                                  bool                           prealigned,
+                                  cudaStream_t                   stream);
 
 }  // namespace nvMolKit
